+++ octomap/src/simple_example.cpp
@@ -59,7 +59,7 @@ int main(int argc, char** argv) {
   for (int x=-20; x<20; x++) {
     for (int y=-20; y<20; y++) {
       for (int z=-20; z<20; z++) {
-        point3d endpoint ((float) x*0.05f, (float) y*0.05f, (float) z*0.05f);
+        point3d endpoint (x*0.05, y*0.05, z*0.05);
         tree.updateNode(endpoint, true); // integrate 'occupied' measurement
       }
     }
@@ -70,7 +70,7 @@ int main(int argc, char** argv) {
   for (int x=-30; x<30; x++) {
     for (int y=-30; y<30; y++) {
       for (int z=-30; z<30; z++) {
-        point3d endpoint ((float) x*0.02f-1.0f, (float) y*0.02f-1.0f, (float) z*0.02f-1.0f);
+        point3d endpoint (x*0.02-1.0, y*0.02-1.0, z*0.02-1.0);
         tree.updateNode(endpoint, false);  // integrate 'free' measurement
       }
     }
